+++ src/main/java/org/example/a/ReservaController.java
@@ -7,23 +7,8 @@ public class ReservaController {
     public ReservaController() {
     }
 
-    public static void verReserva(){
-
-    }
-
-    public static Cliente getCliente() {
-        return cliente;
-    }
-
-    public static void setCliente(Cliente cliente) {
-        ReservaController.cliente = cliente;
-    }
-
     public static Reserva getReserva() {
         return reserva;
     }
 
-    public static void setReserva(Reserva reserva) {
-        ReservaController.reserva = reserva;
-    }
 }
